// src/app/models/weather-response.models.ts
export interface WeatherData {
 coord: Coord
 weather: Weather[]
 base: string
 main: Main
 visibility: number
 wind: Wind
 rain: Rain
 clouds: Clouds
 dt: number
 sys: Sys
 timezone: number
 id: number
 name: string
 cod: number
 
 temp_celcius: string
 temp_min: string
 temp_max: string
 temp_feels_like: string
}

export interface Coord {
 lon: number
 lat: number
}

export interface Weather {
 id: number
 main: string
 description: string
 icon: string
}

export interface Main {
 temp: number
 feels_like: number
 temp_min: number
 temp_max: number
 pressure: number
 humidity: number
 sea_level: number
 grnd_level: number
}

export interface Wind {
 speed: number
 deg: number
 gust: number
}

export interface Rain {
 "1h": number
}

export interface Clouds {
 all: number
}

export interface Sys {
 country: string
 sunrise: number
 sunset: number
}